fix: Target-temperature mis-represented with a configured temperature-factor and hot cache.
When a device was configured with a target temperature-factor other then 1, the target temperature would be mis-represented if you set a target temperature, force-close the home-app and re-open the home app.
The target temperature-factor would wrongly be applied to the previously configured temperature again.

The device cache should at all times represent the data that the tuya API could return, this commit makes the target temperature follow this priciniple.